pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "flask-login>=0.6.3",
    "google-auth>=2.40.3",
    "httpx>=0.28.1",
    "oauthlib>=3.2.2",
    "pydantic>=2.11.5",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
